extracting out runmat-server-client from cli crate
