+++ Cargo.toml
@@ -21,6 +21,7 @@ members = [
     "crates/runmat-logging",
     "crates/runmat-lsp",
     "crates/runmat-cli",
+    "crates/runmat-server-client",
     "crates/runmat-core",
     "crates/runmat-time",
     "crates/runmat-wasm",
@@ -64,6 +65,7 @@ runmat-lexer = { version = "=0.3.2", path = "crates/runmat-lexer" }
 runmat-parser = { version = "=0.3.2", path = "crates/runmat-parser" }
 runmat-hir = { version = "=0.3.2", path = "crates/runmat-hir" }
 runmat-ignition = { version = "=0.3.2", path = "crates/runmat-ignition" }
+runmat-vm = { version = "=0.3.2", path = "crates/runmat-vm" }
 runmat-async = { version = "=0.3.2", path = "crates/runmat-async" }
 runmat-telemetry = { version = "=0.3.2", path = "crates/runmat-telemetry" }
 runmat-core = { version = "=0.3.2", path = "crates/runmat-core", default-features = false }
@@ -79,6 +81,7 @@ runmat-snapshot = { version = "=0.3.2", path = "crates/runmat-snapshot", default
 runmat-filesystem = { version = "=0.3.2", path = "crates/runmat-filesystem" }
 runmat-config = { version = "=0.3.2", path = "crates/runmat-config" }
 runmat-static-analysis = { version = "=0.3.2", path = "crates/runmat-static-analysis" }
+runmat-server-client = { version = "=0.3.2", path = "crates/runmat-server-client" }
 runmat-accelerate = { version = "=0.3.2", path = "crates/runmat-accelerate" }
 runmat-accelerate-api = { version = "=0.3.2", path = "crates/runmat-accelerate-api" }
 runmat-gc-api = { version = "=0.3.2", path = "crates/runmat-gc-api" }
