+++ crates/runmat-server-client/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "runmat-server-client"
+version = "0.3.2"
+edition = "2021"
+license-file = "../../LICENSE.md"
+homepage = "https://runmat.com"
+repository = "https://github.com/runmat-org/runmat"
+
+[dependencies]
+anyhow = { workspace = true }
+base64 = "0.22"
+chrono = { workspace = true }
+dirs = "5.0"
+keyring = "2.3"
+rand = "0.8"
+progenitor-client = "0.11"
+regress = "0.10"
+reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls"] }
+serde = { workspace = true }
+serde_json = { workspace = true }
+sha2 = "0.10"
+tokio = { workspace = true, features = ["net", "time"] }
+url = "2.5"
+uuid = { workspace = true, features = ["serde"] }
+webbrowser = "0.8"
+
+[build-dependencies]
+openapiv3 = "2.0"
+progenitor = "0.11"
+serde_yaml = "0.9"
