+++ crates/runmat-server-client/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod auth;
+pub mod public_api;
